Don't render Mentions component for single user site
In order to enable this when using the site selector, the editor’s
focus event is used as the trigger to either mount the component, or
unmount it (if previously mounted) for a single user site.

In order to enable this when going directly into the editor, the
editor’s init event runs the same logic as above.